requirements.txt
streamlit==1.22.0
pandas==1.5.3
plotly==5.14.1
numpy==1.24.3
Pillow==9.5.0